requirements.txt
bleach==3.1.0
certifi==2018.8.24
chardet==3.0.4
colorama==0.4.3
colorlog==4.1.0
dj-database-url==0.5.0
Django==2.2.2
django-bleach==0.5.0
django-cors-headers==2.4.0
django-js-asset==1.1.0
django-mptt==0.9.1
djangorestframework==3.8.2
factory-boy==2.11.1
Faker==0.9.2
gunicorn==20.0.4
idna==2.7
numpy==1.15.4
pipdeptree==0.13.2
psycopg2==2.8.4
psycopg2-binary==2.8.4
python-dateutil==2.7.5
pytz==2018.5
requests==2.20.1
simplejson==3.16.0
six==1.11.0
sqlparse==0.3.1
text-unidecode==1.2
uritemplate==3.0.0
urllib3==1.23
webencodings==0.5.1
whitenoise==5.1.0
